Move network id fn to circuit and lint
