+++ crates/evm/src/query.rs
@@ -2111,7 +2111,7 @@ fn get_pending_block_env<C: sov_modules_api::Context>(
         latest_block.header.gas_limit,
         latest_block.header.base_fee_per_gas.unwrap_or_default(),
         cfg.base_fee_params,
-        citrea_spec_id
+        citrea_spec_id,
     );
 
     let evm_spec_id = citrea_spec_id_to_evm_spec_id(citrea_spec_id);
