+++ crates/light-client-prover/Cargo.toml
@@ -12,7 +12,7 @@ repository.workspace = true
 # Citrea Deps
 bitcoin-da = { path = "../bitcoin-da", optional = true }
 citrea-common = { path = "../common", optional = true }
-citrea-primitives = { path = "../primitives" }
+citrea-primitives = { path = "../primitives", optional = true }
 citrea-risc0-batch-proof = { path = "../../guests/risc0/batch-proof" }
 prover-services = { path = "../prover-services", optional = true }
 
@@ -59,6 +59,7 @@ native = [
   "dep:sov-mock-da",
   "dep:bitcoin-da",
   "dep:prover-services",
+  "dep:citrea-primitives",
   "dep:citrea-common",
   "dep:sov-prover-storage-manager",
   "dep:sov-db",
