+++ crates/primitives/src/lib.rs
@@ -3,7 +3,6 @@ pub mod compression;
 mod constants;
 pub mod forks;
 pub mod merkle;
-pub mod network;
 pub mod types;
 
 pub use constants::*;
