+++ crates/sequencer/src/mempool.rs
@@ -5,7 +5,6 @@ use alloy_primitives::TxHash;
 use citrea_common::SequencerMempoolConfig;
 use citrea_evm::SYSTEM_SIGNER;
 use citrea_primitives::min_base_fee_per_gas;
-use sov_modules_api::SpecId;
 use reth_execution_types::ChangedAccount;
 use reth_tasks::TaskExecutor;
 use reth_transaction_pool::blobstore::NoopBlobStore;
@@ -15,6 +14,7 @@ use reth_transaction_pool::{
     EthTransactionValidator, Pool, PoolConfig, PoolResult, PoolTransaction, SubPoolLimit,
     TransactionPool, TransactionPoolExt, TransactionValidationTaskExecutor, ValidPoolTransaction,
 };
+use sov_modules_api::SpecId;
 
 pub use crate::db_provider::DbProvider;
